Update Mana fabricator to have a default cost of 1,000,000 RF similarly to the other fabricators.
This fixes issue #63

This will not resolve the issue with existing installs unless the config file is regenerated, but will prevent the exploit on any new installs.

Prior to this change the ratio of energy consumed by the Mana Fabricator vs. the energy produced by the Hippie Dynamo for that Mana was 1 to 4, after this change it is 12.5 to 1. This is more in line with the ratio of the Essentia Dynamo which is 15.6 to 1.