Improve Blood Dynamo fuel values
The Blood Dynamo underproduces power per fuel compared to the other dynamos.

This commit increases the fuel produced per milibucket of blood to be on par with the other Dynamos. This brings the ratio of Blood produced by a Blood fabricator to the amount of Blood consumed to produce that power to 15.6, which is in line with the Essentia Dynamo.