More work to isolate mod code from each other.
This breaks the majority of the mod specific initialization code into Singleton modules. This eliminates any references to mod specific code in common static methods. Referencing a class in a static method causes resolution when the class containing the static method is loaded. This causes that class to be loaded and resolved even if not used. In the case of mod items that extend or implement a class from the "parent" mod, this causes an exception if the parent mod is not loaded.

For example, this change will allow TileEssentiaContainer to extend TileJarFillable directly, but not exception if Technomancy is loaded without Thaumcraft being available.

It also removes some classes that were not being used, as well as some other code cleanup.