Adds the Technoturges Scepter.
This item is a crafting scepter with a vis storage of 150, that can be charged using RF, or a wand recharging station. It is available with both Thaumium and Voidmetal caps.

It also functions as many different types of wrenches, has all of the advanced functionality of EnderIO's Yeta wrench, can rotate blocks just like Thermal Expansion Crescent Hammer, will show you the area of effect of MFR machines, and functions as a Railcraft crowbar.

You must add the src\api\java folder as a source folder in your IDE.